# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API epesantren pegawai",
        "description": "Dokumentasi API epesantren pegawai",
        "contact": {
            "email": "fp3175723@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/absen": {
            "post": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Absen ke hadiran",
                "description": "Mengirim data absensi dengan jenis absensi 'Datang' ",
                "operationId": "18dece7c7fee4337ecfa0515d936f5cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "jenis_absen",
                                    "foto"
                                ],
                                "properties": {
                                    "jenis_absen": {
                                        "type": "string",
                                        "enum": [
                                            "Datang"
                                        ]
                                    },
                                    "foto": {
                                        "description": "Unggah foto bukti absensi",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "catatan_absen": {
                                        "description": "Catatan absensi",
                                        "type": "string"
                                    },
                                    "area_absen": {
                                        "description": "id_area absen",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data absensi berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi data absensi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak dapat absen karena sudah SAKIT pada hari ini."
                                        },
                                        "error": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid atau kedaluwarsa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/pulang": {
            "post": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Absen Pulang",
                "description": "Mengirim data absensi dengan jenis absensi 'Pulang' ",
                "operationId": "c2359c5d502ae78a0b949fe62a63f0ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "jenis_absen",
                                    "foto"
                                ],
                                "properties": {
                                    "jenis_absen": {
                                        "type": "string",
                                        "enum": [
                                            "Pulang"
                                        ],
                                        "example": "PULANG"
                                    },
                                    "foto": {
                                        "description": "Unggah foto bukti absensi",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "catatan_absen": {
                                        "description": "Catatan absensi",
                                        "type": "string"
                                    },
                                    "area_absen": {
                                        "description": "area absensi",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data absensi berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi data absensi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}